[16.0][IMP] budget_appropriation: merge portal and department modules
Consolidate budget_appropriation_portal and budget_appropriation_department
into the main budget_appropriation module to reduce module fragmentation.

Changes:
- Add hr and portal dependencies
- Add department_id field with employee auto-default and READONLY_STATES
- Add portal controllers and templates for preview URL
- Add F4 report data files and QWeb templates
- Add account_ids computed field for hierarchy traversal
- Update budget_move_vals() to pass department_id
- Add button_box with Preview button, hide internal preview page
- Update dependent modules to use budget_appropriation directly

Removed modules:
- budget_appropriation_portal
- budget_appropriation_department
