+++ budget_appropriation/__init__.py
@@ -1,2 +1,3 @@
 from . import models
 from . import utils
+from . import controllers
