--- budget_appropriation_department/__manifest__.py
@@ -1,14 +1,0 @@
-# -*- coding: utf-8 -*-
-{
-    "name": "Budget Appropriation Department",
-    "version": "16.0.1.0.0",
-    "category": "KMITL/Budgeting",
-    "author": "Aginix Technologies",
-    "website": "https://github.com/aginix/kmitl",
-    "depends": ["budget_appropriation", "hr"],
-    "data": ["views/budget_appropriation_views.xml"],
-    "application": False,
-    "installable": True,
-    "auto_install": False,
-    "license": "LGPL-3",
-}
