--- budget_appropriation_department/i18n/th.po
@@ -1,28 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* budget_appropriation_department
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-10-14 09:35+0000\n"
-"PO-Revision-Date: 2025-10-14 16:36+0700\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"Language: th\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=1; plural=0;\n"
-"X-Generator: Poedit 3.7\n"
-
-#. module: budget_appropriation_department
-#: model:ir.model,name:budget_appropriation_department.model_budget_appropriation
-msgid "Budget Appropriation"
-msgstr "จัดสรรงบประมาณ"
-
-#. module: budget_appropriation_department
-#: model:ir.model.fields,field_description:budget_appropriation_department.field_budget_appropriation__department_id
-msgid "Department"
-msgstr "หน่วยงาน"
