--- budget_appropriation_portal/__manifest__.py
@@ -1,20 +1,0 @@
-# -*- coding: utf-8 -*-
-{
-    "name": "Budget_appropriation_portal",
-    "version": "16.0.1.0.0",
-    "summary": """ Budget_appropriation_portal Summary """,
-    "author": "",
-    "website": "",
-    "category": "",
-    "depends": ["budget_appropriation"],
-    "data": [
-        "data/budget_appropriation_f4_action.xml",
-        "data/budget_appropriation_f4_pdf_report.xml",
-        "views/budget_appropriation_views.xml",
-        "views/templates.xml",
-        "report/budget_appropriation_f4_report_templates.xml"
-    ],
-    "installable": True,
-    "auto_install": False,
-    "license": "LGPL-3",
-}
