--- budget_appropriation_portal/models/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import budget_appropriation
