--- budget_appropriation_portal/views/budget_appropriation_views.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-    <!-- View budget.appropriation form -->
-    <record id="view_budget_appropriation_form" model="ir.ui.view">
-        <field name="name">view.budget.appropriation.form</field>
-        <field name="model">budget.appropriation</field>
-        <field name="inherit_id" ref="budget_appropriation.view_budget_appropriation_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//div[hasclass('oe_title')]" position="before">
-                <div class="oe_button_box" name="button_box">
-                    <button
-                        class="oe_stat_button"
-                        type="object"
-                        name="open_record_url"
-                        icon="fa-external-link"
-                        string="Preview"
-                    />
-                </div>
-            </xpath>
-            <xpath expr="//page[@name='preview']" position="attributes">
-                <attribute name="invisible">1</attribute>
-            </xpath>
-        </field>
-    </record>
-
-</odoo>
